pyproject.toml
[project]
name = "data-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.33.0",
    "openpyxl",
    "pandas",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
data-extraction = "data_extraction.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_extraction"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
